G_calendar_module/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
firebase-admin
